Refactor embedding info
