+++ frontend/src/metabase/admin/settings/components/widgets/EmbeddingCustomizationWidget/index.ts
@@ -1,0 +1,1 @@
+export { EmbeddingCustomizationWidget } from "./EmbeddingCustomizationWidget";
